pyproject.toml
[tool.poetry]
name = "sp-spoty"
version = "0.0.1"
description = "SP Spots & Alerts"
authors = [
	"Dariusz Rybi <jiivanq@gmail.com>",
]

[tool.poetry.dependencies]
python = ">=3.10"
prettytable = "=3.9.0"
python-dateutil = "^2.8.2"
requests = "^2.31.0"
